Modularize buildShows function
