+++ src/assets/stylesheets/style.css
@@ -65,7 +65,7 @@ ul {
 .shows-list {
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
-  width: 80%;
+  width: 90%;
   margin: 2rem auto;
 }
